Extract and expose drawBlobOnCanvas() and canvasToBlob() functions
